refactor for clarity
