+++ py/dynesty/dynamicsampler.py
@@ -171,38 +171,6 @@ def weight_function(results, args=None, return_weights=False):
     return (logl_min, logl_max)
 
 
-def _get_update_interval_ratio(update_interval, sample, bound, ndim, nlive,
-                               slices, walks):
-    """
-    Get the update_interval divided by the number of live points
-    """
-
-    # TODO this needs to be updated
-    if update_interval is None:
-        if sample == 'unif':
-            update_interval_frac = 1.5
-        elif sample == 'rwalk':
-            update_interval_frac = 0.15 * walks
-        elif sample == 'slice':
-            update_interval_frac = 0.9 * ndim * slices
-        elif sample == 'rslice':
-            update_interval_frac = 2.0 * slices
-        else:
-            update_interval_frac = 1
-            warnings.warn(
-                "No update_interval set with unknown sampling method: "
-                f"'{sample}'. Defaulting to no 1 update per nlive points.")
-    elif isinstance(update_interval, float):
-        update_interval_frac = update_interval
-    elif isinstance(update_interval, int):
-        update_interval_frac = update_interval * 1. / nlive
-    else:
-        raise RuntimeError(f'Strange update_interval value {update_interval}')
-    if bound == 'none':
-        update_interval_frac = np.inf
-    return update_interval_frac
-
-
 def stopping_function(results,
                       args=None,
                       rstate=None,
@@ -722,7 +690,7 @@ class DynamicSampler:
         # bounding/sampling
         self.bounding = bound
         self.sampling = method
-        self.update_interval_ratio = update_interval_ratio
+        self.bound_update_interval_ratio = update_interval_ratio
         self.first_update = first_update
 
         # internal sampler object
@@ -839,7 +807,7 @@ class DynamicSampler:
             if isinstance(update_interval, float):
                 cur_update_interval_ratio = update_interval
             elif update_interval is None:
-                cur_update_interval_ratio = self.update_interval_ratio
+                cur_update_interval_ratio = self.bound_update_interval_ratio
             else:
                 raise RuntimeError(
                     str.format('Weird update_interval value {}',
